fix classification display
