+++ static/js/components/SimilarSources.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import PropTypes from "prop-types";
-import { useDispatch } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 import Tooltip from "@mui/material/Tooltip";
 import { Link } from "react-router-dom";
 
@@ -8,35 +8,40 @@ import * as summaryActions from "../ducks/summary";
 
 const SimilarSources = ({ source, min_score = 0.9, k = 3 }) => {
   const dispatch = useDispatch();
+  const usePinecone = useSelector((state) => state.config.usePinecone);
   const [simSourceList, setSimSourceList] = useState([]);
 
   useEffect(() => {
-    let tmpList = [];
-    const queryBundle = {
-      objID: source.id,
-      // get an extra source to account for the source itself
-      k: k + 1,
-    };
-    dispatch(summaryActions.fetchSummaryQuery(queryBundle)).then((response) => {
-      if (response.status === "success") {
-        if (response.data) {
-          tmpList = response.data?.query_results;
-          if (tmpList.length > 0) {
-            // remove the source itself from the list
-            tmpList = tmpList.filter((item) => item.id !== source.id);
-            // remove any sources with a score below the threshold
-            tmpList = tmpList.filter((item) => item.score >= min_score);
-            setSimSourceList(tmpList);
-          } else {
-            setSimSourceList([]);
+    if (source?.id && usePinecone) {
+      let tmpList = [];
+      const queryBundle = {
+        objID: source.id,
+        // get an extra source to account for the source itself
+        k: k + 1,
+      };
+      dispatch(summaryActions.fetchSummaryQuery(queryBundle)).then(
+        (response) => {
+          if (response.status === "success") {
+            if (response.data) {
+              tmpList = response.data?.query_results;
+              if (tmpList.length > 0) {
+                // remove the source itself from the list
+                tmpList = tmpList.filter((item) => item.id !== source.id);
+                // remove any sources with a score below the threshold
+                tmpList = tmpList.filter((item) => item.score >= min_score);
+                setSimSourceList(tmpList);
+              } else {
+                setSimSourceList([]);
+              }
+            } else {
+              setSimSourceList([]);
+            }
           }
-        } else {
-          setSimSourceList([]);
-        }
-      }
-      // Don't show an error if the query fails, just don't show any similar sources
-    });
-  }, [dispatch, source, k, min_score]);
+          // Don't show an error if the query fails, just don't show any similar sources
+        },
+      );
+    }
+  }, [dispatch, source, k, min_score, usePinecone]);
 
   return (
     <>
