+++ static/js/components/UpdateSourceSummary.jsx
@@ -22,7 +22,7 @@ const useStyles = makeStyles(() => ({
     margin: "1rem",
   },
   editIcon: {
-    height: "0.75rem",
+    height: "1rem",
     cursor: "pointer",
   },
 }));
